+++ src/source/mod.rs
@@ -85,9 +85,9 @@ mod uniform;
 mod zero;
 
 #[cfg(feature = "noise")]
-mod noise;
+pub mod noise;
 #[cfg(feature = "noise")]
-pub use self::noise::{pink, white, PinkNoise, WhiteNoise};
+pub use self::noise::{Pink, WhiteUniform};
 
 /// A source of samples.
 ///
@@ -698,16 +698,12 @@ impl fmt::Display for SeekError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             SeekError::NotSupported { underlying_source } => {
-                write!(
-                    f,
-                    "Seeking is not supported by source: {}",
-                    underlying_source
-                )
+                write!(f, "Seeking is not supported by source: {underlying_source}")
             }
             #[cfg(feature = "symphonia")]
-            SeekError::SymphoniaDecoder(err) => write!(f, "Error seeking: {}", err),
+            SeekError::SymphoniaDecoder(err) => write!(f, "Error seeking: {err}"),
             #[cfg(feature = "hound")]
-            SeekError::HoundDecoder(err) => write!(f, "Error seeking in wav source: {}", err),
+            SeekError::HoundDecoder(err) => write!(f, "Error seeking in wav source: {err}"),
             SeekError::Other(_) => write!(f, "An error occurred"),
         }
     }